NTFS: read the 64-bit VCN fields at their full width

+++ src/main/java/org/jnode/fs/ntfs/index/IndexBlock.java
@@ -63,7 +63,7 @@ final class IndexBlock extends NTFSRecord {
      * @return
      */
     public long getIndexBlockVCN() {
-        return getUInt32(0x10);
+        return getInt64(0x10);
     }
 
     /**
